comparison-tool/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "migeval"
version = "2.0.0"
description = "Migration health evaluation tool — evaluates migration attempts on absolute quality"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "click>=8.0",
    "pyyaml>=6.0",
    "claude-agent-sdk>=0.1.54",
    "openai>=2.30.0",
]

[project.optional-dependencies]
llm = ["openai>=1.0"]
playwright = ["playwright>=1.40"]
dev = [
    "mypy>=1.8",
    "types-PyYAML",
    "pytest>=8.0",
]

[project.scripts]
migeval = "migeval.cli:main"

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "openai",
    "claude_agent_sdk",
    "claude_agent_sdk.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"migeval/bootstrap.py" = ["E501"]
"migeval/layers/llm_review.py" = ["E501"]
"migeval/layers/runtime.py" = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["migeval"]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "ruff>=0.15.9",
    "types-pyyaml>=6.0.12.20250915",
]
